kernel: spin tbf header to its own file
process.rs was getting a little hard to parse, and this paves the way
for the code to be shared with elf2tab and maybe tockloader-rs someday!
